api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
starlette>=0.36.0
httpx>=0.27.0
rich>=13.0.0
